@@ -1,7 +1,7 @@
 package vaku
 
 import (
-	"github.com/pkg/errors"
+	"fmt"
 )
 
 // folderDestroyWorkerInput takes input/output channels for input to the job
@@ -21,14 +21,14 @@ func (c *Client) FolderDestroy(i *PathInput) error {
 		TrimPathPrefix: false,
 	})
 	if err != nil {
-		return errors.Wrapf(err, "Failed to list %s", i.Path)
+		return fmt.Errorf("failed to list %s: %w", i.Path, err)
 	}
 
 	// Init the path
 	i.opType = "destroy"
 	err = c.InitPathInput(i)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to init path %s", i.Path)
+		return fmt.Errorf("failed to init path %s: %w", i.Path, err)
 	}
 
 	// Concurrency channels for workers
@@ -58,7 +58,7 @@ func (c *Client) FolderDestroy(i *PathInput) error {
 	for j := 0; j < len(list); j++ {
 		o := <-resultsC
 		if o != nil {
-			err = errors.Wrap(o, "Failed to destroy path")
+			err = fmt.Errorf("failed to destroy path: %w", o)
 		}
 	}
 
@@ -73,7 +73,7 @@ func (c *Client) folderDestroyWorker(i *folderDestroyWorkerInput) {
 		if more {
 			err = c.PathDestroy(path)
 			if err != nil {
-				i.resultsC <- errors.Wrapf(err, "Failed to destroy path %s", path.Path)
+				i.resultsC <- fmt.Errorf("failed to destroy path %s: %w", path.Path, err)
 				continue
 			}
 			i.resultsC <- nil
