@@ -1,8 +1,6 @@
 package vaku
 
-import (
-	"github.com/pkg/errors"
-)
+import "fmt"
 
 // PathMove calls PathCopy() with the same inputs followed by PathDelete() on
 // the source if the copy was successful. Note that this will overwrite any existing
@@ -13,13 +11,13 @@ func (c *Client) PathMove(s *PathInput, t *PathInput) error {
 	// Copy the data to the new path
 	err = c.PathCopy(s, t)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to copy data from %s to %s", s.Path, t.Path)
+		return fmt.Errorf("failed to copy data from %s to %s: %w", s.Path, t.Path, err)
 	}
 
 	// Delete the data at the old path
 	err = c.PathDelete(s)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to delete source path %s. This means that the path was copied instead of deleted", s.Path)
+		return fmt.Errorf("failed to delete source path %s. This means that the path was copied instead of deleted: %w", s.Path, err)
 	}
 
 	return err
