@@ -1,8 +1,6 @@
 package vaku
 
-import (
-	"github.com/pkg/errors"
-)
+import "fmt"
 
 // PathUpdate takes in a path with existing data and new data to write to that path.
 // It then merges the data at the existing path with the new data, with precedence given
@@ -13,7 +11,7 @@ func (c *Client) PathUpdate(i *PathInput, d map[string]interface{}) error {
 	// Get old data
 	read, err := c.PathRead(i)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to read data at path %s. PathUpdate only works on existing data", i.Path)
+		return fmt.Errorf("failed to read data at path %s. PathUpdate only works on existing data: %w", i.Path, err)
 	}
 
 	// Generate the new data to write
@@ -24,7 +22,7 @@ func (c *Client) PathUpdate(i *PathInput, d map[string]interface{}) error {
 	// Write the updated data back to vault
 	err = c.PathWrite(i, read)
 	if err != nil {
-		return errors.Wrapf(err, "Failed to write updated data back to %s", i.opPath)
+		return fmt.Errorf("failed to write updated data back to %s: %w", i.opPath, err)
 	}
 
 	return err
