@@ -2,5 +2,5 @@ package vaku
 
 // Version returns the current Vaku API version
 func Version() string {
-	return "1.1.1"
+	return "1.4.1"
 }
